Update logger properties

+++ sbk-api/src/main/java/io/sbk/api/Print.java
@@ -24,5 +24,5 @@ public interface Print {
      * @param percentiles Array of percentiles.
      */
     void print(long bytes, long records, double recsPerSec, double mbPerSec, double avgLatency,
-               int maxLatency, long invalid, long lowerDiscard, long higherDiscard, int[] percentiles);
+               long maxLatency, long invalid, long lowerDiscard, long higherDiscard, long[] percentiles);
 }
